arm64: dts: sunxi: enable USB controller for the Banana Pi M64
Although the Banana Pi M64 shares the same SoC as the Pine64, the USB
configuration is somewhat different: The host only PHY port is connected
to a soldered USB hub, which connects to the two Type-A receptables on
the board. The OTG port is connected to a proper microB socket, so we
can rely on the automatically controlled OTG mode here.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>