arm64: dts: allwinner: enable RTL8723CS for Pinebook
Pinebook features a RTL8723CS SDIO Wi-Fi card on mmc1 bus.

Enable its MMC controller.

As it do not come with an EEPROM storing its MAC address, a device node
of the card is added and set as alias ethernet0, thus MAC address will
be filled by U-Boot for it.

Signed-off-by: Icenowy Zheng <icenowy@aosc.xyz>